# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Swagger Gentcmen API documentation",
        "contact": {
            "email": "nikitosnov@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/client/coupons": {
            "get": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Fetch available coupons",
                "operationId": "Fetch available coupons",
                "parameters": [
                    {
                        "name": "not_expired",
                        "in": "query",
                        "description": "Fecth not expired records",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "with_trashed",
                        "in": "query",
                        "description": "Fecth only trashed items",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "only_trashed",
                        "in": "query",
                        "description": "Fecth only trashed items",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Coupons data fetched successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
        "/api/v1/client/coupons/{coupon}": {
            "get": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Fetch coupon by id",
                "operationId": "Fetch coupon by id",
                "parameters": [
                    {
                        "name": "coupon",
                        "in": "path",
                        "description": "Coupon id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Coupon fetched successfully"
                    },
                    "404": {
                        "description": "Coupon not found"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
        "/api/v1/admin/coupons": {
            "post": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Store a coupon",
                "operationId": "Store a coupon",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CouponVirtualBody"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Coupon created successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Remove coupons",
                "operationId": "Remove coupons",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DestroySomeItemsVirtualBody"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Coupons removed successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/coupons/{coupon}": {
            "put": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Update a coupon",
                "operationId": "Update a coupon",
                "parameters": [
                    {
                        "name": "coupon",
                        "in": "path",
                        "description": "Coupon id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CouponVirtualBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Coupon updated successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    },
                    "404": {
                        "description": "Coupon not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Delete a coupon",
                "operationId": "Delete a coupon",
                "parameters": [
                    {
                        "name": "coupon",
                        "in": "path",
                        "description": "Coupon id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Coupon deleted successfully"
                    },
                    "404": {
                        "description": "Coupon not found"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/admin/coupons/restore/{id}": {
            "post": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Restore a coupon",
                "operationId": "Restore a coupon",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Coupon id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Coupon restored successfully"
                    },
                    "404": {
                        "description": "Coupon not found"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/admin/coupons/soft-delete/permanent-delete": {
            "delete": {
                "tags": [
                    "Coupons"
                ],
                "summary": "Remove permanent coupons",
                "operationId": "Remove permanent coupons",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DestroySomeItemsVirtualBody"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Coupons removed successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CouponVirtualBody": {
                "title": "Create coupon example",
                "description": "Create coupon body data",
                "required": [
                    "name",
                    "cost",
                    "discount"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the coupon",
                        "type": "string",
                        "example": "some name"
                    },
                    "cost": {
                        "title": "cost",
                        "description": "Cost of the coupon",
                        "type": "integer",
                        "example": 450
                    },
                    "discount": {
                        "title": "discount",
                        "description": "discount of the coupon",
                        "type": "integer",
                        "example": 20
                    },
                    "expires_at": {
                        "title": "expires_at",
                        "description": "expires of the coupon",
                        "type": "string",
                        "format": "date"
                    },
                    "entity_type": {
                        "title": "entity_type",
                        "description": "entity type",
                        "type": "string",
                        "example": "Product|ProductCategory"
                    },
                    "entity_id": {
                        "title": "entity_id",
                        "description": "entity id",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "DestroySomeItemsVirtualBody": {
                "title": "Destroy some items example",
                "description": "Destroy some items body data",
                "required": [
                    "ids"
                ],
                "properties": {
                    "ids": {
                        "title": "ids",
                        "description": "Identifiers that should be deleted",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2,
                            3
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}